Add files via upload
This project revolves around the Indian Electoral System, which manages and organizes electoral
data. This database is designed for managing elections, where different constituencies have voters,
and voters participate in elections by casting their votes for candidates. Each candidate is associated
with a specific party and stands for election in a particular constituency. The database allows the
recording of voters, candidates, voting records, and election details.
The primary entities are Constituency, Party, Voter, Election, Candidate, and Voting_Record.
Relationships among these entities are mapped to ensure data integrity, such as a voter being
associated with only one constituency and votes being cast in only one election for a particular
candidate.